caption-model-api/requirements.txt
flask==2.3.3
transformers==4.41.2
torch==2.4.0
pillow==10.4.0
peft==0.12.0
requests==2.32.3
einops==0.8.0
timm==1.0.7
flask-cors==3.0.10
keybert==0.7.0